move getInitialData into own folder
