+++ src/app/routes/radio/getInitialData/index.js
@@ -1,5 +1,5 @@
-import fetchPageData from '../utils/fetchPageData';
-import addIdsToBlocks from './transformers/addIdsToBlocks';
+import fetchPageData from '../../utils/fetchPageData';
+import addIdsToBlocks from '../transformers/addIdsToBlocks';
 
 export default async path => {
   const { json, ...rest } = await fetchPageData(path);
